refactor: consolidate ApiService
