+++ frontend/src/app/components/tv-lightweight/tv-lightweight.component.ts
@@ -1,7 +1,7 @@
 import { Component, ElementRef, OnDestroy, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { createChart, ColorType } from 'lightweight-charts';
-import { ApiService } from '../../services/api.service';
+import { ApiService } from '../../core/services/api.service';
 import { WsService } from '../../services/ws.service';
 import { firstValueFrom } from 'rxjs';
