+++ frontend/src/app/pages/strategies.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { ApiService } from '../services/api.service';
+import { ApiService } from '../core/services/api.service';
 
 @Component({
   standalone: true,
